refactor: use `useModal()`
